ComputerEmployee

+++ Bangazon_Workforce_Management/Controllers/ComputersController.cs
@@ -138,15 +138,25 @@ namespace Bangazon_Workforce_Management.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var computer = await _context.Computer.SingleOrDefaultAsync(m => m.ComputerID == id);
-            _context.Computer.Remove(computer);
-            await _context.SaveChangesAsync();
+            /* var computer = await _context.Computer.SingleOrDefaultAsync(m => m.ComputerID == id);
+             _context.Computer.Remove(computer);
+             await _context.SaveChangesAsync();
+             return RedirectToAction("Index");*/
+            var computer = _context.Computer.SingleOrDefault(m => m.ComputerID == id);
+            var smt = _context.ComputerEmployee.Any(c => c.ComputerID == id);
+
+            if (!smt)
+            {
+               _context.Computer.Remove(computer);
+                await _context.SaveChangesAsync();
+            }
+           
             return RedirectToAction("Index");
         }
 
-        private bool ComputerExists(int id)
+        private bool ComputerExists(int CompID)
         {
-            return _context.Computer.Any(e => e.ComputerID == id);
+            return _context.Computer.Any(e => e.ComputerID == CompID);
         }
     }
 }
